+++ api/specs/webserver/scripts/openapi_nih_sparc.py
@@ -15,7 +15,7 @@ from models_library.generics import Envelope
 from models_library.projects import ProjectID
 from models_library.services import ServiceKey, ServiceKeyVersion
 from pydantic import HttpUrl, PositiveInt
-from simcore_service_webserver.studies_dispatcher._handlers_rest import (
+from simcore_service_webserver.studies_dispatcher._rest_handlers import (
     ServiceGet,
     Viewer,
 )
