+++ services/web/server/src/simcore_service_webserver/studies_dispatcher/_redirects_handlers.py
