+++ services/web/server/tests/unit/isolated/test_studies_dispatcher_core.py
@@ -19,13 +19,13 @@ from pydantic.main import BaseModel
 from pydantic.networks import HttpUrl
 from pytest_simcore.helpers.utils_services import list_fake_file_consumers
 from servicelib.aiohttp.requests_validation import parse_request_query_parameters_as
-from simcore_service_webserver.studies_dispatcher._handlers_redirects import (
-    RedirectionQueryParams,
-)
 from simcore_service_webserver.studies_dispatcher._projects import (
     UserInfo,
     ViewerInfo,
     _create_project_with_filepicker_and_service,
+)
+from simcore_service_webserver.studies_dispatcher._redirects_handlers import (
+    RedirectionQueryParams,
 )
 from yarl import URL
